Day10-19/Day12/source-code/Azure-AI-Agent-Remote-MCP-main/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
azure-identity==1.15.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
azure-ai-projects==1.0.0b3
azure-ai-agents==1.0.0b3